requirements.txt
django
djangorestframework
djangorestframework_simplejwt
channels
pymongo
drf-yasg
beautifulsoup4